+++ sdk/core/core-rest-pipeline/src/createPipelineFromOptions.ts
@@ -65,10 +65,10 @@ export function createPipelineFromOptions(options: InternalPipelineOptions): Pip
   }
 
   pipeline.addPolicy(formDataPolicy());
-  pipeline.addPolicy(tracingPolicy(options.userAgentOptions));
   pipeline.addPolicy(userAgentPolicy(options.userAgentOptions));
   pipeline.addPolicy(setClientRequestIdPolicy());
   pipeline.addPolicy(defaultRetryPolicy(), { phase: "Retry" });
+  pipeline.addPolicy(tracingPolicy(options.userAgentOptions), { phase: "Retry" });
   pipeline.addPolicy(redirectPolicy(options.redirectOptions), { afterPhase: "Retry" });
   pipeline.addPolicy(logPolicy(options.loggingOptions), { afterPhase: "Retry" });