+++ sdk/core/core-rest-pipeline/CHANGELOG.md
@@ -9,8 +9,9 @@
 - A new `retryPolicy` centralizes the retry logic and allows adding retry strategies to any pipeline. With it, we're exposing some new types:
   - `RetryStrategy` defines whether to retry and how to retry.
   - `RetryStrategyState` keeps track of the last retry and controls how to do the subsequent retries.
-- Previous retry policies have been enhanced with better error handling, logging and tracing.
+- Previous retry policies have been enhanced with better error handling.
 - A new `defaultRetryPolicy` is added, which has the same behavior as all the other retry policies combined (`throttlingRetryPolicy`, `systemErrorRetryPolicy` and `exponentialRetryPolicy`). 
+- `createPipelineFromOptions` has been updated to reflect retries on the tracing spans.
 
 ### Breaking Changes